JST4-93 Implemented modal popup for non-logged in users
