+++ docker-compose.yml
@@ -25,7 +25,7 @@ services:
     ports:
       - 5000:5000
     environment:
-      DB_HOST: db
+      DB_HOST: ${DB_HOST}
       DB_USER: ${DB_USER}
       DB_PASSWORD: ${DB_PASSWORD}
       DB_NAME: ${DB_NAME}
@@ -45,7 +45,7 @@ services:
     ports:
       - 3000:3000
     volumes:
-    - ./webUI/src:/app/src
+      - ./webUI/src:/app/src
 
   admin:
     container_name: admin
