+++ webUI/src/components/common/modal/modal.scss
@@ -1,0 +1,37 @@
+.modal {
+  height: 100vh;
+  width: 100vw;
+  background-color: rgba(0, 0, 0, 0.2);
+  position: fixed;
+  top: 0;
+  left: 0;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  opacity: 0;
+  transition: 0.5s all;
+  pointer-events: none;
+}
+.modal.active {
+  opacity: 1;
+  z-index: 10;
+  pointer-events: all;
+}
+.modal__content {
+  padding: 40px 20px;
+  border-radius: 6px;
+  transition: 0.5s all;
+  background-color: white;
+  transform: scale(0.5);
+  width: 50vw;
+  text-align: center;
+}
+.modal__content.active {
+  opacity: 1;
+  z-index: 10;
+  pointer-events: all;
+  transform: scale(1);
+}
+.title {
+  text-align: center;
+}
